Renamed package conv to masking
