+++ src/main/java/dev/dokan/dokan_java/constants/dokany/MountError.java
@@ -1,6 +1,6 @@
 package dev.dokan.dokan_java.constants.dokany;
 
-import dev.dokan.dokan_java.conv.EnumInteger;
+import dev.dokan.dokan_java.masking.EnumInteger;
 
 /**
  * Return values of com.dokan.java.NativeMethods#DokanMain(DokanOptions, DokanOperations)
