+++ src/main/java/dev/dokan/dokan_java/constants/dokany/MountOption.java
@@ -1,7 +1,7 @@
 package dev.dokan.dokan_java.constants.dokany;
 
-import dev.dokan.dokan_java.conv.MaskValueSet;
-import dev.dokan.dokan_java.conv.MaskValueEnum;
+import dev.dokan.dokan_java.masking.MaskValueSet;
+import dev.dokan.dokan_java.masking.MaskValueEnum;
 import dev.dokan.dokan_java.structure.DokanOptions;
 
 /**
