+++ src/main/java/dev/dokan/dokan_java/constants/microsoft/CreateDisposition.java
@@ -2,7 +2,7 @@ package dev.dokan.dokan_java.constants.microsoft;
 
 import dev.dokan.dokan_java.DokanNativeMethods;
 import dev.dokan.dokan_java.DokanOperations;
-import dev.dokan.dokan_java.conv.EnumInteger;
+import dev.dokan.dokan_java.masking.EnumInteger;
 import com.sun.jna.ptr.IntByReference;
 
 /**
