+++ src/main/java/dev/dokan/dokan_java/examples/DirListingFileSystem.java
@@ -13,8 +13,8 @@ import dev.dokan.dokan_java.FileSystemInformation;
 import dev.dokan.dokan_java.constants.microsoft.CreateDisposition;
 import dev.dokan.dokan_java.constants.microsoft.CreateOption;
 import dev.dokan.dokan_java.constants.microsoft.NtStatuses;
-import dev.dokan.dokan_java.conv.EnumInteger;
-import dev.dokan.dokan_java.conv.MaskValueSet;
+import dev.dokan.dokan_java.masking.EnumInteger;
+import dev.dokan.dokan_java.masking.MaskValueSet;
 import dev.dokan.dokan_java.structure.ByHandleFileInformation;
 import dev.dokan.dokan_java.structure.DokanFileInfo;
 import dev.dokan.dokan_java.structure.DokanIOSecurityContext;
