+++ src/main/java/dev/dokan/dokan_java/masking/MaskValueEnum.java
@@ -1,4 +1,4 @@
-package dev.dokan.dokan_java.conv;
+package dev.dokan.dokan_java.masking;
 
 
 public interface MaskValueEnum extends EnumInteger {
