+++ src/test/java/dev/dokan/dokan_java/masking/EnumIntegerTest.java
@@ -1,4 +1,4 @@
-package dev.dokan.dokan_java.conv;
+package dev.dokan.dokan_java.masking;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
